added validations and auth flow

+++ src/controllers/userControllers.js
@@ -1,5 +1,6 @@
 const { hash, compare } = require('bcryptjs')
 const knex = require('../database/knex/connection')
+const AppError = require('../utils/AppError')
 
 class UserController {
     
@@ -8,7 +9,13 @@ class UserController {
         
         const alreadyExists = await knex('users').where({ email }).first()
         if (alreadyExists) {
-            return res.status(400).json({ message: 'Este email já está em uso' })
+            throw new AppError('Este email já está em uso', 401)
+        }
+
+        const validationPasswordLength = password.length <= 8 || password.length > 30
+
+        if (validationPasswordLength) {
+            throw new AppError("Senha deve ter no mínimo 8 caracteres e, no máximo, 30 caracteres", 401)
         }
         
         const hashed = await hash(password, 8)
@@ -18,9 +25,15 @@ class UserController {
     }
 
     async delete(req, res) {
-        const { id } = req.params
+        const user_id = req.user.id
+
+        const user = await knex('users').where({ id: user_id }).first()
+
+        if (!user) {
+            throw new AppError('Usuário inexistente', 404)
+        }
         
-        await knex('users').where({ id }).delete()
+        await knex('users').where({ id: user_id }).delete()
 
         res.status(200).json({ message: 'Conta excluída com sucesso' })
     }
@@ -32,19 +45,37 @@ class UserController {
     }
 
     async update(req, res) {
-        const { name, email, password, oldPassword} = req.body
-        const { id } = req.params
+        const { name, email, newPassword, oldPassword, city } = req.body
+        const user_id = req.user.id
 
-        const user = await knex('users').where({ id }).first()
+        const user = await knex('users').where({ id: user_id }).first()
         const matchPassword = await compare(oldPassword, user.password)
-
+        const passwordAlreadyBeingUsed = await compare(newPassword, user.password)
+        
+        user.name = name ?? user.name
+        user.email = email ?? user.email
+        user.city = city ?? user.city
+        
         if(!matchPassword) {
-            return res.status(400).json({ message: 'A senha antiga não confere' })
+            throw new AppError("Senha antiga incorreta", 403)
         }
 
-        const hashed = await hash(password, 8)
+        if (email) {
+            const isEmailInUse = await knex('users').where({ email }).first()
+            
+            if (isEmailInUse) {
+                throw new AppError(`E-mail ${email} já está em uso`, 401)
+            }
+        }
+            
 
-        await knex('users').where({ id }).update({ name, email, password: hashed })
+        if (passwordAlreadyBeingUsed) {
+            throw new AppError("Nova senha não pode ser igual à anterior!", 400)
+        }
+
+        user.password = await hash(newPassword, 8)
+
+        await knex('users').where({ id: user_id }).update(user)
 
         res.status(200).json({ message: 'Dados atualizados com sucesso' })
